Eos management network: added store for management, data network and DNS setting

See merge request eos/csm!229

+++ gui/src/services/api.ts
@@ -19,54 +19,66 @@ import router from "../router"; // Get router object from our router.ts
 // Set valid token into each request header
 // Note: - if welcome page no need of auth token
 //       - if create admin user page no need of auth token
-axios.interceptors.request.use((config) => {
+axios.interceptors.request.use(
+  config => {
     const conststr = require("../common/const-string.json");
     const token = localStorage.getItem(conststr.access_token);
     if (token) {
-        config.headers.Authorization = token;
+      config.headers.Authorization = token;
     }
 
     return config;
-}, (error) => {
+  },
+  error => {
     Promise.reject(error);
-});
+  }
+);
 
 // Handle response
-axios.interceptors.response.use((response) => {
-
+axios.interceptors.response.use(
+  response => {
     return response;
-}, (error) => {
+  },
+  error => {
     // Handle Unauthorised response. Re-route to login page if unauthorised response received.
-    if (error.response.status === 401) {
-        const conststr = require("../common/const-string.json");
-        localStorage.removeItem(conststr.access_token);
-        router.push("/login");
+    if (error.response && error.response.status === 401) {
+      const conststr = require("../common/const-string.json");
+      localStorage.removeItem(conststr.access_token);
+      router.push("/login");
 
-        return Promise.reject(error);
+      return Promise.reject(error);
     }
-});
-
+  }
+);
 
 export abstract class Api {
-
-    // Wrapper method to for get api
-    public static async getAll(url: string, queryParams?: object) {
-        return await axios.get(url, { params: queryParams });
+  // Wrapper method to for get api
+  public static async getAll(url: string, queryParams?: object) {
+    return await axios.get(url, { params: queryParams });
+  }
+  // Wrapper method for update api
+  public static async patch(url: string, payload: object, id: number) {
+    if (!id) {
+      return;
     }
-    // Wrapper method for update api
-    public static async patch(url: string, payload: object, id: number) {
-        return await axios.patch(url + "/" + id, payload);
+    return await axios.patch(url + "/" + id, payload);
+  }
+  // Wrapper method for post api
+  public static async post(url: string, payload: object) {
+    return await axios.post(url, payload);
+  }
+  // Wrapper method for post api
+  public static async delete(url: string, id: string) {
+    if (!id) {
+      return;
     }
-    // Wrapper method for post api
-    public static async post(url: string, payload: object) {
-        return await axios.post(url, payload);
+    return await axios.delete(url + "/" + id);
+  }
+  // Wrapper method for update api
+  public static async put(url: string, payload: object, id: string) {
+    if (!id) {
+      return;
     }
-    // Wrapper method for post api
-    public static async delete(url: string, id: string) {
-        return await axios.delete(url + "/" + id);
-    }
-    // Wrapper method for update api
-    public static async put(url: string, payload: object, id: string) {
-        return await axios.put(url + "/" + id, payload);
-    }
+    return await axios.put(url + "/" + id, payload);
+  }
 }
